Don't report changes from None to 'none' (=> None)
Some qvm-prefs properties take 'none' and interpret it as None.
This state module compares the None to 'none' and (kinda rightly)
reports a difference. What it doesn't realize though is that
qvm-prefs turns 'none' into None for certain properties.

To be on the safe side, only do this for those specific properties
which already do 'none' => None, and require other (future) ones
to be added here too.