Fix handling pci_strictreset in qvm.prefs
The interaction between pci_strictreset and pcidevs is really obscure,
and in fact there were several issues with its handling. The one
prompting this changes is inability to change the value for already
attached devices. That's because "pcidevs" value did not change, so the
whole loop (that would change the option) was skipped. But even then,
there the condition to change the value was negated (compared
"pci_strictreset" to "no-strict-reset" without negating one of them).
While at it, fix other issues found there.

Fix several issues:
 - distinguish between pci_strictreset not included in the state and set
   to True - this allows removing the no-strict-reset option too
 - report when pci_strictreset value is changed
 - fix changing the value for already attached devices
 - setting pci_strictreset works only together with 'pcidevs' - report
   missing 'pcidevs' as an error

Fixes QubesOS/qubes-issues#8514