new device API for salt
It is now possible for `pcidevs` to refer to devices in the following ways: `*:dev_id`, `port_id:*`, `port_id:dev_id`, or `port_id`. The last format creates a mapping `port_id:dev_id`, where `dev_id` is automatically retrieved from the device found in the `port_id` port. If the device is not found, Salt will raise an error.
This is compatible with the `qvm-pci` tool.